Fixes crash introduced in r254 (download refactoring).
git-svn-id: http://svn.flexget.com/trunk@255 3942dd89-8c5d-46d7-aeed-044bccf3e60c

+++ modules/output_download.py
@@ -146,9 +146,10 @@ class ModuleDownload:
                 feed.fail(entry)
                 log.exception('Error while writing: %s' % e)
             # remove temp file if it remains due exceptions
-            if os.path.exists(entry['file']):
-                log.debug('removing temp file %s (left behind)' % entry['file'])
-                os.remove(entry['file'])
+            if entry.has_key('file'):
+                if os.path.exists(entry['file']):
+                    log.debug('removing temp file %s (left behind)' % entry['file'])
+                    os.remove(entry['file'])
 
     def output(self, feed, entry):
         """Moves temp-file into final destination"""
